+++ lib/src/field/advanced_multi_select_field_controller.dart
@@ -4,14 +4,18 @@ import 'package:leancode_forms/src/field/advanced_field_controller.dart';
 class AdvancedMultiSelectFieldController<V, E extends Object>
     extends AdvancedFieldController<Set<V>, E> {
   /// Creates a new [AdvancedMultiSelectFieldController].
+  ///
+  /// Both [initialValue] and [options] are copied, so mutating the collections
+  /// you pass in never reaches this field.
   AdvancedMultiSelectFieldController({
-    required super.initialValue,
+    required Set<V> initialValue,
     super.validator,
     super.asyncValidation,
     super.focusNode,
-    required this.options,
+    required List<V> options,
     super.name,
-  });
+  })  : options = List.of(options),
+        super(initialValue: Set.of(initialValue));
 
   /// List of options to select from.
   final List<V> options;